pyproject.toml
[project]
name = "implement-guardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
